Split out utils file into multiple files
