+++ src/utils/compose-reducers.js
@@ -1,0 +1,4 @@
+export default function composeReducers(reducers) {
+  return (state, action) => reducers.reduce((prevState, reducer) =>
+    reducer(prevState, action), state);
+}
