+++ src/utils/request-statuses.js
@@ -1,0 +1,8 @@
+// These are statuses for in-flight requests. If a request has no status
+// associated with it, then it would have a status of `NULL`.
+export default {
+  PENDING: 'PENDING',
+  SUCCEEDED: 'SUCCEEDED',
+  FAILED: 'FAILED',
+  NULL: 'NULL'
+};
