+++ src/utils/update-many-resource-metas.js
@@ -1,0 +1,15 @@
+// Similar to `updateResourceMeta`, but it accepts an array of IDs instead of
+// a single ID. Used for bulk updating meta.
+export default function updateManyResourceMetas({meta, newMeta, ids, replace}) {
+  const next = replace ? {} : {...meta};
+
+  ids.forEach((id) => {
+    const current = next[id];
+    next[id] = {
+      ...current,
+      ...newMeta
+    };
+  });
+
+  return next;
+}
